### 2022-10-26 #4883 @editorialbot
[PRE REVIEW]: ticktack: A Python package for carbon box modelling
**Submitting author:** @benjaminpope (<a href="http://orcid.org/0000-0003-2595-9114">Benjamin Pope</a>)
**Repository:** https://github.com/SharmaLlama/ticktack
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
