### 2022-10-26 #4883 @dfm
@benjaminpope — Thanks for your submission! I'm the managing editor for this submissions JOSS "track", so I'm responsible for assigning an editor to handle your submission. Unfortunately, all of the suitable editors are working at capacity so I'm going to need to waitlist this submission for now, and I'll aim to assign an editor ASAP. Thanks for your patience!
