### 2022-10-26 #4883 @benjaminpope
Hi @dfm,

No trouble at all - end of semester here is very busy too!

I think joshuakt would be a good reviewer, who has done carbon box modelling in an astrobiology-geophysical context; in the archaeological context a lot more of the work is closed-source, but andrewcparnell and nevrome have respectively written stable isotope and radiocarbon analysis codes and could also be good reviewers.

All the best,

Ben
