### 2023-01-17 #5084 @editorialbot
[REVIEW]: ticktack: A Python package for carbon box modelling
**Submitting author:** @benjaminpope (<a href="http://orcid.org/0000-0003-2595-9114">Benjamin Pope</a>)
**Repository:** https://github.com/SharmaLlama/ticktack
**Branch with paper.md** (empty if default branch):
**Version:** 1.1.0
