### 2023-01-17 #5084 @dfm
@benjaminpope — Take a look at [those DOI suggestions from the bot above](https://github.com/openjournals/joss-reviews/issues/5084#issuecomment-1386009930) and please remove any entries from the bibtex that aren't referenced in the text!
