+++ joss/paper.bib
@@ -440,11 +440,16 @@ archivePrefix = {arXiv},
 }
 
 @article{zhang2022,
-Author = {Qingyuan {Zhang} and Utkarsh {Sharma} and Jordan A. {Dennis} and Andrea {Scifo} and Margot {Kuitems} and Ulf {Buentgen} and Mathew J. {Owens} and Michael W. {Dee} and Benjamin J. S. {Pope}},
-Title = {Modelling Cosmic Radiation Events in the Tree-ring Radiocarbon Record},
-Year = {2022},
-Eprint = {arXiv:2210.13775},
-Doi = {10.1098/rspa.2022.0497.},
+  doi = {10.1098/rspa.2022.0497},
+  url = {https://doi.org/10.1098/rspa.2022.0497},
+  year = {2022},
+  month = oct,
+  publisher = {The Royal Society},
+  volume = {478},
+  number = {2266},
+  author = {Qingyuan Zhang and Utkarsh Sharma and Jordan A. Dennis and Andrea Scifo and Margot Kuitems and Ulf B\"{u}ntgen and Mathew J. Owens and Michael W. Dee and Benjamin J. S. Pope},
+  title = {Modelling cosmic radiation events in the tree-ring radiocarbon record},
+  journal = {Proceedings of the Royal Society A: Mathematical,  Physical and Engineering Sciences}
 }
 
 @BOOK{dorman2004,
@@ -827,7 +832,7 @@ volume = {158},
 pages = {164-171},
 year = {2017},
 issn = {0277-3791},
-doi = {https://doi.org/10.1016/j.quascirev.2016.12.024},
+doi = {10.1016/j.quascirev.2016.12.024},
 url = {https://www.sciencedirect.com/science/article/pii/S0277379116305017},
 author = {Clive Oppenheimer and Lukas Wacker and Jiandong Xu and Juan Diego Galván and Markus Stoffel and Sébastien Guillet and Christophe Corona and Michael Sigl and Nicola {Di Cosmo} and Irka Hajdas and Bo Pan and Remco Breuker and Lea Schneider and Jan Esper and Jie Fei and James O.S. Hammond and Ulf Büntgen},
 keywords = {Holocene, Paleoclimatology, Eastern Asia, Cosmogenic isotopes, Ice cores, Tree-rings}
@@ -1694,7 +1699,7 @@ number = {4},
 pages = {321-325},
 year = {1989},
 issn = {0893-9659},
-doi = {https://doi.org/10.1016/0893-9659(89)90079-7},
+doi = {10.1016/0893-9659(89)90079-7},
 url = {https://www.sciencedirect.com/science/article/pii/0893965989900797},
 author = {P. Bogacki and L.F. Shampine}
 }
