### 2023-01-26 #5084 @joshuakt
Hi Dan, thanks for the reminder - I will do this in the next couple of days.

On Thu, Jan 26, 2023 at 9:32 AM Dan Foreman-Mackey ***@***.***>
wrote:

> @joshuakt <https://github.com/joshuakt> — This is a ping to keep this on
> your radar. The first step of generating your checklist as described above,
> and starting on the first few boxes should be pretty quick. Let me know if
> you run into issues, and I'll send an email if we don't hear from you since
> I know GitHub notifications can get lost.
>
