### 2023-01-28 #5084 @joshuakt
Some initial comments on code usability:

I had a few issues trying to get ticktack installed - I kept getting errors when I tried installing in my usual conda environment(s) because I had the wrong version of this or that. In the end, I had to create a new conda environment with most of the requirements from requirements.txt:

conda create -n tictac -c conda-forge python numpy scipy jax jaxlib astropy matplotlib diffrax emcee h5py==3.3.0 pandas seaborn tqdm tinygp pip

conda activate tictac

And then I had to install the remaining three packages (chainconsumer, hdfdict, jax-cosmo) with pip install. Only then did pip install ticktack work properly.

I found it was also necessary to pip install pytest to successfully run the tests in the tests folder.

After installation, I attempted to run joss_figure.py to reproduce the results in the manuscript. It took around 10 minutes to get through the burn-in and "running production" but then I got an error "NameError: name 'ChainConsumer' is not defined"

I attempted to fix this by adding the line "from chainconsumer import ChainConsumer" to joss_figure.py. This seems to work because the code finished running and reproduced Fig. 1 from the manuscript.

The manuscript itself is clear, contains a persuasive statement of need, and sufficient references for the reader to dig into the model details. Finally, I worked through the fitting documentation and everything was explained succinctly and the code ran without issue.
