### 2023-02-01 #5084 @dfm
@benjaminpope — I haven't had any success finding a second reviewer for this submission so I propose that I will do a review myself. The only potential issue with this is that @benjaminpope and I have co-authored two (related) papers within the past 4 years (Nth authors in both cases):

- https://joss.theoj.org/papers/10.21105/joss.02125
- https://iopscience.iop.org/article/10.3847/1538-3881/ab7d38

I don't believe that these make me unable to provide impartial judgement, but I wanted to give @benjaminpope, any other authors, and @arfon (as an external JOSS editor) a chance to chime in with any concerns about this potential perceived conflict of interest.
