### 2023-02-11 SharmaLlama/ticktack#33 @dfm
Hard constraint on h5py version causes installation nightmares
What is the motivation for pinning the required version number of h5py here:

https://github.com/SharmaLlama/ticktack/blob/e5beeaf556a5f3bbb4f15e4c034b2aa4eba66b98/requirements.txt#L9

That's a pretty old version that doesn't have builds on conda or PyPI for Python 3.10, and h5py can be a pain to build from source. I'd recommend loosening that constraint if possible, or explaining clearly in the docs why it's required with some tips for how to install it (e.g. Python<3.10 is required).

ref: https://github.com/openjournals/joss-reviews/issues/5084
