### 2023-02-11 SharmaLlama/ticktack#34 @dfm
Installing from source fails when numpy is not installed
But… numpy is not actually required to build this project. I'd recommend removing the first 3 lines of `setup.py` since they're not actually used!

https://github.com/SharmaLlama/ticktack/blob/e5beeaf556a5f3bbb4f15e4c034b2aa4eba66b98/setup.py#L1-L3

ref: https://github.com/openjournals/joss-reviews/issues/5084
