### 2023-02-11 SharmaLlama/ticktack#36 @dfm
Where do the datasets included in the `data` directory come from
The source, acknowledgement, and appropriate licensing text should be included with the datasets in the `data` directory.

ref: https://github.com/openjournals/joss-reviews/issues/5084
