### 2023-02-11 SharmaLlama/ticktack#37 @dfm
How do we run the tests?
I was able to figure it out because I know how these things work :D, but the docs should include a discussion of the preferred steps for running the unit tests.

ref: https://github.com/openjournals/joss-reviews/issues/5084
