+++ setup.py
@@ -1,6 +1,3 @@
-from numpy.distutils.core import setup, Extension
-from numpy.distutils.misc_util import Configuration
-import distutils.sysconfig as ds
 from setuptools import setup
 import os
 import codecs
