+++ src/data/datasets/REFERENCE_775AD-early-S.txt
@@ -1,0 +1,6 @@
+Büntgen18_DAR01.csv; Büntgen U., et al., 2018, Nature Communications, 9, 3605; https://www.nature.com/articles/s41467-018-06036-0
+Büntgen18_DAR06.csv; Büntgen U., et al., 2018, Nature Communications, 9, 3605; https://www.nature.com/articles/s41467-018-06036-0
+Büntgen18_DAR07.csv; Büntgen U., et al., 2018, Nature Communications, 9, 3605; https://www.nature.com/articles/s41467-018-06036-0
+Büntgen18_PAT02.csv; Büntgen U., et al., 2018, Nature Communications, 9, 3605; https://www.nature.com/articles/s41467-018-06036-0
+Büntgen18_PAT03.csv; Büntgen U., et al., 2018, Nature Communications, 9, 3605; https://www.nature.com/articles/s41467-018-06036-0
+Büntgen18_TAS01.csv; Büntgen U., et al., 2018, Nature Communications, 9, 3605; https://www.nature.com/articles/s41467-018-06036-0
