+++ src/data/datasets/REFERENCE_993AD-N.txt
@@ -1,0 +1,9 @@
+Büntgen18_ALT02.csv; Büntgen U., et al., 2018, Nature Communications, 9, 3605; https://www.nature.com/articles/s41467-018-06036-0
+Büntgen18_CHI01.csv; Büntgen U., et al., 2018, Nature Communications, 9, 3605; https://www.nature.com/articles/s41467-018-06036-0
+Büntgen18_MON05.csv; Büntgen U., et al., 2018, Nature Communications, 9, 3605; https://www.nature.com/articles/s41467-018-06036-0
+Büntgen18_SWE01.csv; Büntgen U., et al., 2018, Nature Communications, 9, 3605; https://www.nature.com/articles/s41467-018-06036-0
+Büntgen18_SWE02.csv; Büntgen U., et al., 2018, Nature Communications, 9, 3605; https://www.nature.com/articles/s41467-018-06036-0
+Büntgen18_SWE03.csv; Büntgen U., et al., 2018, Nature Communications, 9, 3605; https://www.nature.com/articles/s41467-018-06036-0
+Büntgen18_SWE04.csv; Büntgen U., et al., 2018, Nature Communications, 9, 3605; https://www.nature.com/articles/s41467-018-06036-0
+
+
