### 2023-02-24 83d8b44
Documenting the procedure to run the tests (see #37).
