### 2023-02-24 827123b
Revert "Documenting the procedure to run the tests (see #37)."
This reverts commit 83d8b448d1e08ec1c5121235621ae5757552c1ba.
