### 2023-02-24 1b73d61
Documenting the procedure for running the tests (see #37).
