### 2023-02-25 SharmaLlama/ticktack#36 @dfm
Thanks @qingyuanzhang3! I think that the added references are ok, but pretty much the bare minimum. If possible I think it would be good to perhaps add a readme to that directory and/or more discussion of this to the docs.
