### 2023-02-25 SharmaLlama/ticktack#37 @dfm
I'm not sure I would have closed this since as far as I can tell the generated docs haven't actually been updated to include the new text!
