### 2023-02-26 9712c2a
Allowing automatic detection of the tests (see #37).
