### 2023-02-27 SharmaLlama/ticktack#39 @dfm
Temporary files should be removed from repo
The `tests/__pycache__` directory is a temporary directory that should be removed and probably added to the `.gitignore`.

ref: https://github.com/openjournals/joss-reviews/issues/5084
