### 2023-02-27 SharmaLlama/ticktack#40 @dfm
Testing instructions should mention that `pytest` needs to be installed
... and perhaps link to the `pytest` docs.

ref: https://github.com/openjournals/joss-reviews/issues/5084
