### 2023-02-27 SharmaLlama/ticktack#41 @dfm
Community guidelines and contributor guide
One of the JOSS requirements is:

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

I couldn't find this documented on the docs page or in the GitHub repo. This page makes some advice for setting up contributor guides: https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors

ref: https://github.com/openjournals/joss-reviews/issues/5084
