### 2023-02-27 SharmaLlama/ticktack#42 @dfm
API docs could be consolidated 
Right now, the docs page has each class' API docs on a separate top-level page. I found that this leads to a somewhat confusing interface. I'd recommend either putting these onto one page with section headings for each class or otherwise consolidating these pages under a top-level "API docs" directory or similar. This isn't _required_ for the JOSS review, but I think it would substantially improve the user experience.

ref: https://github.com/openjournals/joss-reviews/issues/5084
