### 2023-02-27 SharmaLlama/ticktack#43 @dfm
Inconsistent results in "Fitting multiple datasets using MultiFitter" tutorial
When running the ["Fitting multiple datasets using MultiFitter"](https://sharmallama.github.io/ticktack/notebooks/02_MultiFitter/) tutorial, I find that when I make the final `chain_summary` plot, I get the following warnings:

```python
WARNING:chainconsumer:Parameter start date (yr) in chain Chain 0 is not constrained
WARNING:chainconsumer:Parameter duration (yr) in chain Chain 0 is not constrained
```

And no results are reported in the axis titles:

![82858979-345c-4fd6-bc6a-ba17c20e7992](https://user-images.githubusercontent.com/350282/221685404-4f49c7a1-594f-495a-8797-2a14f4ed97db.png)

The results do seem qualitatively similar, so this doesn't seem like a major issue, but it might be worth double checking that this isn't the result of something bigger!

ref: https://github.com/openjournals/joss-reviews/issues/5084
