+++ CONTRIBUTING.md
@@ -1,0 +1,23 @@
+# Contributing to ticktack
+
+We encourage anyone interested in improving `ticktack` to contribute to the code and documentation! We hope that this is only the beginning of a wider project bringing fast, differentiable carbon box modelling to radiocarbon science.
+
+## Getting Started
+
+We encourage contributors who are new to open-source development to consult the [AstroPy docs](https://docs.astropy.org/en/latest/development/workflow/development_workflow.html) for its suggested workflow for contributions.
+
+## Project Scope
+
+`ticktack` is focused on carbon box modelling for analysing tree-ring radiocarbon time series. At present, we are soliciting contributions that improve its capabilities within this scope, or that modestly extend the scope - but it is not well suited to modelling the carbon cycle in general terms including kinetics or general circulation models.
+
+## Questions and Support
+
+If you have questions about `ticktack` or encounter problems when using it, please open an [Issue](https://github.com/SharmaLlama/ticktack/issues) on the GitHub issue tracker. If you see an Issue you can help solve, please comment!
+
+## Suggestions and Discussions
+
+If you have ideas for extensions to `ticktack`, including expansions to its scope or use case or ways to substantially refactor and improve the code, please go to [Discussions](https://github.com/SharmaLlama/ticktack/discussions) on GitHub and create a thread.
+
+## Code of Conduct
+
+There is a [Code of Conduct](https://github.com/SharmaLlama/ticktack/blob/main/CODE_OF_CONDUCT.md) that we expect all contributors to `ticktack` to adhere to.
