### 2023-03-01 SharmaLlama/ticktack#41 @benjaminpope
[Added](https://github.com/SharmaLlama/ticktack/blob/main/CONTRIBUTING.md).
