### 2023-03-01 SharmaLlama/ticktack#43 @benjaminpope
Ah, well spotted. You see above that we added in a couple of lines

```python
samples = sampler.copy()
samples[:,1] = 10**samples[:,1] # duration not log duration
samples[:,-1] = 10**samples[:,-1] # area not log area
```

but didn't copy this at the bottom, and evidently didn't re-run the cell before committing the docs. This is because we switched to default parametrization of log duration and log spike production, rather than real-space. This is now fixed.
