### 2023-03-01 #5084 @benjaminpope
Hi @dfm,

My apologies for taking so long - last week was the first week of the teaching semester and things got on top of me a little.

Thank you both for your helpful and insightful comments and engagement with the code, and for carefully identifying issues with the initial submission.

We have now addressed and closed all the issues raised by @dfm. We have also fixed both the install issue and the `joss_figure.py` issue identified by @joshuakt. I want to explicitly acknowledge @sharmallama, @qingyuanzhang3, and @jordan-dennis for their work building this package and addressing the issues raised in review.

We are now happy to resubmit and continue the review.

All the best,

Ben
