### 2023-03-08 #5084 @dfm
@benjaminpope — Besides @Jordan-Dennis' docs point, I think the ball is back to me an @joshuakt to finish going through our checklists. JOSS reviews are typically iterative so I expect that we'll have some more feedback as we go through the last steps. I'll aim to get back to it ASAP, and @joshuakt can you take a look again soon too? Thanks!
