### 2023-03-09 SharmaLlama/ticktack#44 @dfm
Tests raise various deprecation warnings
Some of these are potential future bugs!

1. In `fitting.py` all of the label names with `\Delta` in them should be either protected with a double slash (i.e. `\\Delta`), or use a raw string (i.e. `r"$\Delta^{14}$"`)
<summary>warnings</summary>

```
../src/ticktack/fitting.py:309
  /Users/dforemanmackey/src/joss/sharmallama/ticktack/src/ticktack/fitting.py:309: DeprecationWarning: invalid escape sequence '\D'
    ax1.set_ylabel("$\Delta^{14}$C (‰)")

../src/ticktack/fitting.py:316
  /Users/dforemanmackey/src/joss/sharmallama/ticktack/src/ticktack/fitting.py:316: DeprecationWarning: invalid escape sequence '\D'
    markersize=6.5, elinewidth=3, label="average $\Delta^{14}$C")

../src/ticktack/fitting.py:327
  /Users/dforemanmackey/src/joss/sharmallama/ticktack/src/ticktack/fitting.py:327: DeprecationWarning: invalid escape sequence '\D'
    custom_lines.append(Line2D([0], [0], color="k", marker="o", lw=1.5, label="average $\Delta^{14}$C"))

../src/ticktack/fitting.py:1642
  /Users/dforemanmackey/src/joss/sharmallama/ticktack/src/ticktack/fitting.py:1642: DeprecationWarning: invalid escape sequence '\D'
    elinewidth=elinewidth, label="average $\Delta^{14}$C")

../src/ticktack/fitting.py:1664
  /Users/dforemanmackey/src/joss/sharmallama/ticktack/src/ticktack/fitting.py:1664: DeprecationWarning: invalid escape sequence '\D'
    custom_lines.append(Line2D([0], [0], color="k", marker="o", lw=1.5, label="average $\Delta^{14}$C"))

../src/ticktack/fitting.py:1666
  /Users/dforemanmackey/src/joss/sharmallama/ticktack/src/ticktack/fitting.py:1666: DeprecationWarning: invalid escape sequence '\D'
    ax1.set_ylabel("$\Delta^{14}$C (‰)")

../src/ticktack/fitting.py:1751
  /Users/dforemanmackey/src/joss/sharmallama/ticktack/src/ticktack/fitting.py:1751: DeprecationWarning: invalid escape sequence '\D'
    markersize=markersize, elinewidth=elinewidth, label="average $\Delta^{14}$C")

../src/ticktack/fitting.py:1774
  /Users/dforemanmackey/src/joss/sharmallama/ticktack/src/ticktack/fitting.py:1774: DeprecationWarning: invalid escape sequence '\D'
    custom_lines.append(Line2D([0], [0], color="k", marker="o", lw=1.5, label="average $\Delta^{14}$C"))

../src/ticktack/fitting.py:1776
  /Users/dforemanmackey/src/joss/sharmallama/ticktack/src/ticktack/fitting.py:1776: DeprecationWarning: invalid escape sequence '\D'
    ax1.set_ylabel("$\Delta^{14}$C (‰)")
```

2. The use of the `seaborn` styles in `matplotlib` is deprecated (and I generally think it's extremely bad form for a library to change the plotting style on import like this!!).
<summary>warnings</summary>

```
../src/ticktack/fitting.py:32
  /Users/dforemanmackey/src/joss/sharmallama/ticktack/src/ticktack/fitting.py:32: MatplotlibDeprecationWarning: The seaborn styles shipped by Matplotlib are deprecated since 3.6, as they no longer correspond to the styles shipped by seaborn. However, they will remain available as 'seaborn-v0_8-<style>'. Alternatively, directly use the seaborn API instead.
    mpl.style.use('seaborn-colorblind')
```

3. I'm not sure what the story is with this last one, but it's probably worth looking into:
<summary>warnings</summary>

```
tests/test_fitting.py::test_correlation_plot
  /Users/dforemanmackey/src/joss/sharmallama/ticktack/src/ticktack/fitting.py:184: MatplotlibDeprecationWarning: Auto-removal of overlapping axes is deprecated since 3.6 and will be removed two minor releases later; explicitly call ax.remove() as needed.
    ax = plt.subplot(plot_grid[:, :-1])
```

Ref: https://github.com/openjournals/joss-reviews/issues/5084
