### 2023-03-09 #5084 @dfm
I've checked all but 2 of my review items:

- I think that the documentation/licensing for the included data is still insufficient: https://github.com/SharmaLlama/ticktack/issues/36
- The tests emit some deprecation warnings that can be easily fixed (there are also some caused by dependencies, but we'll have to live with those!): https://github.com/SharmaLlama/ticktack/issues/44
