### 2023-03-12 970dc11
raw strings for delta
