### 2023-03-12 SharmaLlama/ticktack#44 @benjaminpope
Hi @dfm,

Good point about the seaborn-colorblind - this shouldn't have been there and we have removed it.

The strings containing \Delta are now all raw strings.

There is obscure syntax about the correlation plot warning and I think this is now fixed.

The only warning I get in unit tests now is a deprecation warning from the ChainConsumer dependency, which we can't do much about.

Closing this issue.
