+++ src/ticktack/data/README.md
@@ -1,0 +1,15 @@
+# Datasets
+
+This repository comes packaged with all datasets used in the [Zhang et al, 2022](https://arxiv.org/abs/2210.13775) first science paper using `ticktack`. These data are publicly available, and necessary to reproduce figures in the aforementioned paper and the accompanying Sharma et al JOSS paper.
+
+Under `src/ticktack/data/datasets/` these are listed, by event, as `.csv` files.
+
+## References
+
+Datsets have filenames labelled by provenance:
+
+- `Buntgen18*`: the COSMIC sample, from [Büntgen U., et al., 2018, Nature Communications, 9, 3605](https://www.nature.com/articles/s41467-018-06036-0)
+- `Scifo19*`: [Scifo A., et al., 2019, Scientific Reports, 9, 17056](https://www.nature.com/articles/s41598-019-53296-x)
+- `Sakurai20*`: [Sakurai H., et al., 2020](https://www.nature.com/articles/s41598-019-57273-2)
+- `Brehm21*`: are from [Brehm N., et al., 2021, Nature Geoscience, 14, 10](https://www.nature.com/articles/s41561-020-00674-0).
+- `Miyake21*`: [Miyake F., et al., 2021, Geophys. Res. Lett., 48, e93419](https://agupubs.onlinelibrary.wiley.com/doi/10.1029/2021GL093419)
