### 2023-03-12 SharmaLlama/ticktack#36 @benjaminpope
This has now been added to a docs page `docs.md` and to `src/ticktack/data/README.md` as requested.
