### 2023-03-14 #5084 @dfm
@benjaminpope — Can you check those missing DOIs, and make sure that only references that are actually cited are included in the `.bib` file? It looks to me like there are some references that are included in the `.bib` file, but not cited. Thanks!
