### 2023-03-16 7616ce8
update dois - only schulman book missing
