+++ joss/paper.bib
@@ -3,10 +3,10 @@
 
 @misc{scipy, title = {{SciPy}: Open source scientific tools for Python}, url = {http://www.scipy.org/}, author = {Jones, E. and Oliphant, T. and Peterson, P. and Others}, year = {2001} }
 
-@Article{matplotlib, Author = {Hunter, J. D.}, Title = {Matplotlib: A 2D graphics environment}, Journal = {Computing In Science \& Engineering}, Volume = {9}, Number = {3}, Pages = {90--95}, abstract = {Matplotlib is a 2D graphics package used for Python for application development, interactive scripting, and publication-quality image generation across user interfaces and operating systems.}, publisher = {IEEE COMPUTER SOC}, year = 2007 }
+@Article{matplotlib, Author = {Hunter, J. D.}, Title = {Matplotlib: A 2D graphics environment}, Journal = {Computing In Science \& Engineering}, Volume = {9}, Number = {3}, Pages = {90--95}, abstract = {Matplotlib is a 2D graphics package used for Python for application development, interactive scripting, and publication-quality image generation across user interfaces and operating systems.}, publisher = {IEEE COMPUTER SOC}, year = 2007, doi={10.1109/mcse.2007.55} }
 
 
-@article{numpy, title={The NumPy array: a structure for efficient numerical computation}, author={Van Der Walt, Stefan and Colbert, S Chris and Varoquaux, Gael}, journal={Computing in Science \& Engineering}, volume={13}, number={2}, pages={22--30}, year={2011}, publisher={AIP Publishing} }
+@article{numpy, title={The NumPy array: a structure for efficient numerical computation}, author={Van Der Walt, Stefan and Colbert, S Chris and Varoquaux, Gael}, journal={Computing in Science \& Engineering}, volume={13}, number={2}, pages={22--30}, year={2011}, publisher={AIP Publishing},doi={10.1109/mcse.2011.37}}
 
 @software{jax,
   author = {James Bradbury and Roy Frostig and Peter Hawkins and Matthew James Johnson and Chris Leary and Dougal Maclaurin and Skye Wanderman-Milne},
@@ -1068,21 +1068,26 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@ARTICLE{luhn22,
+@ARTICLE{luhn2022,
        author = {{Luhn}, Jacob K. and {Wright}, Jason T. and {Henry}, Gregory W. and {Saar}, Steven H. and {Baum}, Anna C.},
         title = "{HD 166620: Portrait of a Star Entering a Grand Magnetic Minimum}",
-      journal = {arXiv e-prints},
-     keywords = {Astrophysics - Solar and Stellar Astrophysics},
+      journal = {\apjl},
+     keywords = {Stellar activity, Maunder minimum, Solar cycle, 1580, 1015, 1487, Astrophysics - Solar and Stellar Astrophysics},
          year = 2022,
-        month = jul,
-          eid = {arXiv:2207.00612},
-        pages = {arXiv:2207.00612},
+        month = sep,
+       volume = {936},
+       number = {2},
+          eid = {L23},
+        pages = {L23},
+          doi = {10.3847/2041-8213/ac8b13},
 archivePrefix = {arXiv},
        eprint = {2207.00612},
  primaryClass = {astro-ph.SR},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2022arXiv220700612L},
[…]
