+++ joss/paper.bib
@@ -1,13 +1,4 @@
 
-@Article{ipython, Author = {P\'erez, Fernando and Granger, Brian E.}, Title = {{IP}ython: a System for Interactive Scientific Computing}, Journal = {Computing in Science and Engineering}, Volume = {9}, Number = {3}, Pages = {21--29}, month = may, year = 2007, url = "http://ipython.org", ISSN = "1521-9615", doi = {10.1109/MCSE.2007.53}, publisher = {IEEE Computer Society}, }
-
-@misc{scipy, title = {{SciPy}: Open source scientific tools for Python}, url = {http://www.scipy.org/}, author = {Jones, E. and Oliphant, T. and Peterson, P. and Others}, year = {2001} }
-
-@Article{matplotlib, Author = {Hunter, J. D.}, Title = {Matplotlib: A 2D graphics environment}, Journal = {Computing In Science \& Engineering}, Volume = {9}, Number = {3}, Pages = {90--95}, abstract = {Matplotlib is a 2D graphics package used for Python for application development, interactive scripting, and publication-quality image generation across user interfaces and operating systems.}, publisher = {IEEE COMPUTER SOC}, year = 2007, doi={10.1109/mcse.2007.55} }
-
-
-@article{numpy, title={The NumPy array: a structure for efficient numerical computation}, author={Van Der Walt, Stefan and Colbert, S Chris and Varoquaux, Gael}, journal={Computing in Science \& Engineering}, volume={13}, number={2}, pages={22--30}, year={2011}, publisher={AIP Publishing},doi={10.1109/mcse.2011.37}}
-
 @software{jax,
   author = {James Bradbury and Roy Frostig and Peter Hawkins and Matthew James Johnson and Chris Leary and Dougal Maclaurin and Skye Wanderman-Milne},
   title = {{JAX}: composable transformations of {P}ython+{N}um{P}y programs},
@@ -16,493 +7,6 @@
   year = {2018},
 }
 
-@ARTICLE{chainconsumer,
-   author = {{Hinton}, S.~R.},
-    title = "{ChainConsumer}",
-  journal = {The Journal of Open Source Software},
-     year = 2016,
-    month = aug,
-   volume = 1,
-      eid = {00045},
-    pages = {00045},
-      doi = {10.21105/joss.00045},
-   adsurl = {http://adsabs.harvard.edu/abs/2016JOSS....1...45H},
-}
-
-@phdthesis{miyakephd,
-  title={Reconstruction of cosmic-ray intensity in the past from measurements of radiocarbon in tree rings},
-  author={{Miyake}, Fusa},
-  year={2013},
-  school={Nagoya University}
-}
-
-@ARTICLE{miyake12,
-       author = {{Miyake}, Fusa and {Nagaya}, Kentaro and {Masuda}, Kimiaki and {Nakamura}, Toshio},
-        title = "{A signature of cosmic-ray increase in AD 774-775 from tree rings in Japan}",
-      journal = {Nature},
-         year = 2012,
-        month = jun,
-       volume = {486},
-       number = {7402},
-        pages = {240-242},
-          doi = {10.1038/nature11123},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2012Natur.486..240M},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
-@ARTICLE{miyake13,
-       author = {{Miyake}, Fusa and {Masuda}, Kimiaki and {Nakamura}, Toshio},
-        title = "{Another rapid event in the carbon-14 content of tree rings}",
-      journal = {Nature Communications},
[…]
