### 2023-03-16 #5084 @benjaminpope
@dfm fixed references - short .bib file (didn't realise this was an issue), DOIs fixed. Thanks.
