+++ src/ticktack/__init__.py
@@ -1,4 +1,4 @@
 name = "ticktack"
-__version__ = "1.0.1"
+__version__ = "1.1.0"
 
 from .ticktack import *
