### 2023-03-20 #5084 @benjaminpope
Merged your typesetting PR.

Incremented version number to 1.1.0.

Ping @sharmallama about Zenodo archive.
