### 2023-03-22 #5084 @dfm
@SharmaLlama — Thanks! For some reason (probably left over from my PRs), I'm still listed as an author on the Zenodo archive. I'm going to proceed with the final processing steps, but can you remove me from the author list ASAP? Thanks!
